Moving cascading model binding code into .Web
