+++ src/Components/Components/test/RouteViewTest.cs
@@ -2,7 +2,6 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 
 using System.Diagnostics.CodeAnalysis;
-using Microsoft.AspNetCore.Components.Binding;
 using Microsoft.AspNetCore.Components.Rendering;
 using Microsoft.AspNetCore.Components.Test.Helpers;
 using Microsoft.Extensions.DependencyInjection;
@@ -21,8 +20,6 @@ public class RouteViewTest
         var serviceCollection = new ServiceCollection();
         _navigationManager = new RouteViewTestNavigationManager();
         serviceCollection.AddSingleton<NavigationManager>(_navigationManager);
-        serviceCollection.AddSingleton<IFormValueSupplier, TestFormValueSupplier>();
-        serviceCollection.AddSingleton<CascadingModelBindingProvider, CascadingQueryModelBindingProvider>();
         var services = serviceCollection.BuildServiceProvider();
         _renderer = new TestRenderer(services);
 
@@ -75,32 +72,15 @@ public class RouteViewTest
             frame => AssertFrame.Component<TestLayout>(frame, subtreeLength: 2, sequence: 0),
             frame => AssertFrame.Attribute(frame, nameof(LayoutComponentBase.Body), sequence: 1));
 
-        // Assert: TestLayout renders cascading model binder
+        // Assert: TestLayout renders page
         var testLayoutComponentId = batch.GetComponentFrames<TestLayout>().Single().ComponentId;
         var testLayoutFrames = _renderer.GetCurrentRenderTreeFrames(testLayoutComponentId).AsEnumerable();
         Assert.Collection(testLayoutFrames,
             frame => AssertFrame.Text(frame, "Layout starts here", sequence: 0),
             frame => AssertFrame.Region(frame, subtreeLength: 3),
-            frame => AssertFrame.Component<CascadingModelBinder>(frame, sequence: 0, subtreeLength: 2),
-            frame => AssertFrame.Attribute(frame, nameof(CascadingModelBinder.ChildContent), typeof(RenderFragment<ModelBindingContext>), sequence: 1),
-            frame => AssertFrame.Text(frame, "Layout ends here", sequence: 2));
-
-        // Assert: Cascading model binder renders CascadingValue<ModelBindingContext>
-        var cascadingModelBinderComponentId = batch.GetComponentFrames<CascadingModelBinder>().Single().ComponentId;
-        var cascadingModelBinderFrames = _renderer.GetCurrentRenderTreeFrames(cascadingModelBinderComponentId).AsEnumerable();
-        Assert.Collection(cascadingModelBinderFrames,
-            frame => AssertFrame.Component<CascadingValue<ModelBindingContext>>(frame, sequence: 0, subtreeLength: 4),
-            frame => AssertFrame.Attribute(frame, nameof(CascadingValue<ModelBindingContext>.IsFixed), false, sequence: 1),
-            frame => AssertFrame.Attribute(frame, nameof(CascadingValue<ModelBindingContext>.Value), typeof(ModelBindingContext), sequence: 2),
-            frame => AssertFrame.Attribute(frame, nameof(CascadingValue<ModelBindingContext>.ChildContent), typeof(RenderFragment), sequence: 3));
-
-        // Assert: CascadingValue<ModelBindingContext> renders page
-        var cascadingValueComponentId = batch.GetComponentFrames<CascadingValue<ModelBindingContext>>().Single().ComponentId;
-        var cascadingValueFrames = _renderer.GetCurrentRenderTreeFrames(cascadingValueComponentId).AsEnumerable();
-        Assert.Collection(cascadingValueFrames,
-            frame => AssertFrame.Region(frame, sequence: 0, subtreeLength: 3),
             frame => AssertFrame.Component<ComponentWithLayout>(frame, sequence: 0, subtreeLength: 2),
-            frame => AssertFrame.Attribute(frame, nameof(ComponentWithLayout.Message), "Test message", sequence: 1));
+            frame => AssertFrame.Attribute(frame, nameof(ComponentWithLayout.Message), "Test message", sequence: 1),
+            frame => AssertFrame.Text(frame, "Layout ends here", sequence: 2));
 
         // Assert: page itself is rendered, having received parameters from the original route data
         var pageComponentId = batch.GetComponentFrames<ComponentWithLayout>().Single().ComponentId;
@@ -109,7 +89,7 @@ public class RouteViewTest
             frame => AssertFrame.Text(frame, "Hello from the page with message 'Test message'", sequence: 0));
 
         // Assert: nothing else was rendered
-        Assert.Equal(6, batch.DiffsInOrder.Count);
+        Assert.Equal(4, batch.DiffsInOrder.Count);
     }
 
     [Fact]
@@ -240,11 +220,5 @@ public class RouteViewTest
             builder.AddContent(1, Body);
             builder.AddContent(2, "OtherLayout ends here");
         }
-    }
-
-    private class TestFormValueSupplier : IFormValueSupplier
-    {
-        public bool CanBind(Type valueType, string formName = null) => false;
-        public void Bind(FormValueSupplierContext context) { }
     }
 }
